OutputHandling.py: Fix __str__ on BasicMorePrinter
We are overriding the default object.__str__ of calling __repr__
because UniSet defines a custom __str__.

A tests that was previously disabled for portability was reenabled,
and we generate the expected result dynamically.
The other test that was disabled for allocation_behaves_as_originally
remain disabled by PORTABLE_TEST because that's more work to reenable,
and it does not relate to this bug.

Fixes #4