horizon.c: Fix segfault during recursive dealloc
During recursive dealloc, if a child is the Horizon heapyc object
it can invoke horizon_remove and cause the unpatching of all the
tp_dealloc. When this returns to parect, such as subtype_dealloc,
it can still refer to the patched version of tp_dealloc,
horizon_patched_dealloc, and invoke it. That would cause a
segfault because rm.types is no longer valid.

We check if rm.types is valid, and if not and tp_dealloc seems
unpatched, call tp_dealloc and don't do dict lookup.

Fixes #10