No more multi-interpreter shenanigans in Py3.9+
Given that the general direction is per-interpreter isolation,
making things work will be harder and harder. It is already almost
impossible for the new-interpreter-in-new-thread to work in Py3.9,
since the new thread cannot send GIL drop request across interp
boundary to the old thread.

The only user of this is currently remote monitor. We can probably
rewrite it so that it is a thread in the same interpreter. Though,
we don't have the interpreter-namespace isolation like we did
before. Kindof a shame but oh well.

Closes #18