RootState: Add traverse, dir, and getattr for _PyInterpreterFrame
No relate yet. Dir and getattr lacks support for localsplus.

Honestly I'm ot sure how to even do relate for localsplus. It needs
both the frame name and the local variable name, which isn't something
path supports. Though, we can add it. But then, ByVia classifier would
be much less useful if a bunch of local variables share the same name.

(Is it possible for, maybe, a pycapsule to represent a
_PyInterpreterFrame?)

For #41