test_Path.py: Fix path ordering when both sources are same size
The paths are sorted in an unexpected way. Test both cases (though
only one case happens per Python version), and document what is
going on behind the test case.

In Python 3.11 both the object and the list have the same size. The
object grew by 8 bytes compared to Python 3.10.

For #41